+++ psycopg/psycopg/server_cursor.py
@@ -172,7 +172,7 @@ class ServerCursorMixin(BaseCursor[ConnectionType, Row]):
         )
         yield from self._conn._exec_command(query)
 
-    def _make_declare_statement(self, query: Query) -> sql.Composable:
+    def _make_declare_statement(self, query: Query) -> sql.Composed:
 
         if isinstance(query, bytes):
             query = query.decode(self._encoding)